feat(websocket):refactor websocket to a file of its own and add comments
